backend/requirements.txt
flask
flask-cors
requests
python-dotenv
mysql-connector-python
beautifulsoup4
openpyxl

aiogram==2.25.1
aiohttp==3.8.6